% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_metric.R
\name{authors_metric}
\alias{authors_metric}
\title{Gets the 'authors metric' for a given filter}
\usage{
authors_metric(code, filter, file = NULL, save = FALSE,
  truncateAt = NULL)
}
\arguments{
\item{code}{The account code to use}

\item{filter}{The filter for the data}

\item{file}{An optional file name to save a CSV file to}

\item{save}{Set to TRUE if you'd like a dialog to choose where to save your CSV}

\item{truncateAt}{Optional number of results - rest will become "Others"}
}
\value{
A tibble of your data
}
\description{
Gets the 'authors metric' for a given filter
}
\examples{

authors_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
