% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{brand_metric}
\alias{brand_metric}
\title{Returns data grouped by brand}
\usage{
brand_metric(code, filter, file = NULL, save = FALSE)
}
\arguments{
\item{code}{The account code to use}

\item{filter}{The filter for the data}

\item{file}{An optional file name to save a CSV file to}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV}
}
\value{
A tibble of your data
}
\description{
Returns data grouped by brand
}
\examples{

brand_metric("QUIR01BA", "published inthelast week  and brand isorchildof 10006")
}
