% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_week.R
\name{day_of_week_metric}
\alias{day_of_week_metric}
\title{Get data for determining the best day of the week to post}
\usage{
day_of_week_metric(code, filter, file = NULL, save = FALSE)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter to use}

\item{file}{A filename to save the data to as a CSV}

\item{save}{Set to true to be given a save file dialog}
}
\value{
A tibble of data
}
\description{
Get data for determining the best day of the week to post
}
\examples{

day_of_week_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")

}
