% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{location_metric}
\alias{location_metric}
\title{Gives information grouped by country}
\usage{
location_metric(code, filter, file = NULL, save = FALSE,
  truncateAt = NULL)
}
\arguments{
\item{code}{The account code to use}

\item{filter}{The filter for the data}

\item{file}{An optional file name to save a CSV file to}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV}

\item{truncateAt}{Optional number of results - rest will become "Others". This takes the
top countries based on the volume of mentions from the site.}
}
\value{
A tibble of your data
}
\description{
Gives information grouped by country
}
\examples{

location_metric("QUIR01BA", "published inthelast week  and brand isorchildof 10006")
}
