% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_metric.R
\name{plot_sentiment_metric}
\alias{plot_sentiment_metric}
\title{Plot the sentiment chart}
\usage{
plot_sentiment_metric(code, filter, useVerified = TRUE)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter for data}

\item{useVerified}{Only use verified data for sentiment - set to TRUE by default}
}
\description{
Plot the sentiment chart
}
\examples{

plot_sentiment_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
