% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day.R
\name{plot_time_of_day_metric}
\alias{plot_time_of_day_metric}
\title{Plots the time of day to post (by hour)}
\usage{
plot_time_of_day_metric(account, filter)
}
\arguments{
\item{account}{An account code}

\item{filter}{A filter for data}
}
\value{
the ggplot object
}
\description{
Plots the time of day to post (by hour)
}
