% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returning_authors.R
\name{returning_authors}
\alias{returning_authors}
\title{Checks whether authors from current filter appeared in a previous filter}
\usage{
returning_authors(code, previousFilter, currentFilter, file = NULL,
  save = FALSE)
}
\arguments{
\item{code}{The account code to use}

\item{previousFilter}{Filter for the time period you want to compare your authors against}

\item{currentFilter}{Filter for the current time period whose authors you want to check}

\item{file}{An optional file name to save a CSV file to}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV}
}
\value{
A tibble of your data
}
\description{
Checks whether authors from current filter appeared in a previous filter
}
