% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_metric.R
\name{sentiment_metric}
\alias{sentiment_metric}
\title{Returns the data for sentiment chart}
\usage{
sentiment_metric(code, filter, file = NULL, save = FALSE)
}
\arguments{
\item{code}{account code}

\item{filter}{A filter for data}

\item{file}{An optional file name to save a CSV file to.}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV.}
}
\value{
A tibble of your data
}
\description{
Returns the data for sentiment chart
}
\examples{
sentiment_metric("QUIR01BA", "published inthelast week")

}
