% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_metric.R
\name{sites_metric}
\alias{sites_metric}
\title{Gets the 'stats metric' for a given filter}
\usage{
sites_metric(code, filter, file = NULL, save = FALSE, truncateAt = NULL)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter for data}

\item{file}{An optional file name to save a CSV file to.}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV.}

\item{truncateAt}{Optional number of results - rest will become "Others". This takes the
top sites based on the volume of mentions from the site.}
}
\value{
A tibble of your data
}
\description{
Gets the 'stats metric' for a given filter
}
\examples{

sites_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
