% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metric.R
\name{stats_metric}
\alias{stats_metric}
\title{Gets the 'stats metric' for a given filter}
\usage{
stats_metric(code, filter, file = NULL, save = FALSE)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter for data}

\item{file}{An optional file name to save a CSV file to.}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV.}
}
\value{
A tibble of your data
}
\description{
Gets the 'stats metric' for a given filter
}
\examples{
stats_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
