% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_brandseye}
\alias{theme_brandseye}
\title{Provides a BrandsEye ggplot2 theme}
\usage{
theme_brandseye()
}
\value{
The theme object.
}
\description{
This is still a work in progress, but should provide a ggplot2 theme
for charts.
}
