% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day.R
\name{time_of_day}
\alias{time_of_day}
\title{Get data for determining time of day to post (by hour)}
\usage{
time_of_day(account, filter)
}
\arguments{
\item{account}{An account code}

\item{filter}{A filter to use}
}
\value{
A tibble of data
}
\description{
Get data for determining time of day to post (by hour)
}
