% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day.R
\name{time_of_day_metric}
\alias{time_of_day_metric}
\title{Get data for determining the best time of day to post (by hour)}
\usage{
time_of_day_metric(code, filter, file = NULL, save = FALSE)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter to use}

\item{file}{A filename to save the data to as a CSV}

\item{save}{Set to true to be given a save file dialog.}
}
\value{
A tibble of data
}
\description{
Get data for determining the best time of day to post (by hour)
}
\examples{

time_of_day_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
