% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-sentiment.R
\name{volume_sentiment}
\alias{volume_sentiment}
\title{Returns the data for creating a volume/sentiment}
\usage{
volume_sentiment(code, filter, group = "day", file = NULL)
}
\arguments{
\item{code}{The account code to use}

\item{filter}{The filter for the data.}

\item{group}{An optional string to group data by. Can be "day", "week", "month"}

\item{file}{An optional file name to save a CSV file to.}
}
\value{
A tibble of your data
}
\description{
Returns the data for creating a volume/sentiment
}
