% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-sentiment.R
\name{volume_sentiment_metric}
\alias{volume_sentiment_metric}
\title{Returns the data for creating volume/sentiment charts}
\usage{
volume_sentiment_metric(code, filter, group = "day", file = NULL,
  save = FALSE)
}
\arguments{
\item{code}{The account code to use}

\item{filter}{The filter for the data}

\item{group}{An optional string to group data by. Can be "day", "week", "month"}

\item{file}{An optional file name to save a CSV file to}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV}
}
\value{
A tibble of your data
}
\description{
Returns the data for creating volume/sentiment charts
}
\examples{

volume_sentiment_metric("QUIR01BA", "published inthelast week  and brand isorchildof 10006")
}
