% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_metric.R
\name{wordcloud_metric}
\alias{wordcloud_metric}
\title{Pulls breakdown of volume by keyword}
\usage{
wordcloud_metric(code, filter, exclude = NULL, save = FALSE, file = NULL)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter for data}

\item{exclude}{A character vector of terms you want to exclude}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV}

\item{file}{An optional file name to save a CSV file to}
}
\value{
Return a list of words and volume of mentions
}
\description{
Pulls breakdown of volume by keyword
}
\examples{

wordcloud_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")

}
