% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metric.R
\name{get_stats}
\alias{get_stats}
\title{Gets the 'stats metric' for a given filter}
\usage{
get_stats(code, filter)
}
\arguments{
\item{code}{account code}

\item{filter}{A filter for data}
}
\value{
A tibble of your data
}
\description{
Gets the 'stats metric' for a given filter
}
\examples{
#' get_stats("QUIR01BA", "published inthelast week")
}
