% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_metric.R
\name{plot_sentiment_metric}
\alias{plot_sentiment_metric}
\title{Plot the sentiment chart}
\usage{
plot_sentiment_metric(code, filter)
}
\arguments{
\item{code}{account code}

\item{filter}{A filter for data}
}
\description{
Plot the sentiment chart
}
\examples{
plot_sentimment_metric("QUIR01BA", "published inthelast week")
}
