% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-sentiment.R
\name{plot_volume_sentiment}
\alias{plot_volume_sentiment}
\title{Plots volume overlayed with sentiment.}
\usage{
plot_volume_sentiment(account, filter, group = "day")
}
\arguments{
\item{account}{An account code}

\item{filter}{A filter for data}

\item{group}{A string indicating how you want your data gruoped.}
}
\value{
the ggplot object
}
\description{
Plots volume overlayed with sentiment.
}
