% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{topics_metric}
\alias{topics_metric}
\title{Gets the Topics for a given filter}
\usage{
topics_metric(code, filter, file = NULL, save = FALSE, truncateAt = NULL,
  showParents = TRUE, showChildren = TRUE)
}
\arguments{
\item{code}{An account code}

\item{filter}{A filter for data}

\item{file}{An optional file name to save a CSV file to.}

\item{save}{Set to TRUE if you'd like a dialog file to choose where to save your CSV.}

\item{truncateAt}{Optional number of results - rest will become "Others". This takes the
top topics based on the volume of mentions from the site.}

\item{showParents}{Set this to FALSE if you don't want any parent topics included in the
results.}

\item{showChildren}{Set this to FALSE if you don't want any child topics included in the
results.}
}
\value{
A tibble of your data
}
\description{
Gets the Topics for a given filter
}
\examples{

\dontrun{
topics_metric("QUIR01BA", "published inthelast week and brand isorchildof 10006")
}
}
